// src/vs/platform/agentHost/node/codex/protocol/generated/v2/ToolRequestUserInputResponse.ts
// AUTOGENERATED - do not edit.
// Source: `codex app-server generate-ts --experimental`
// Generated from @openai/codex 0.142.0, licensed Apache-2.0.
// Modified by VS Code to rewrite imports and apply repository formatting.
// See README.md in this directory for provenance and licensing notes.
// Regenerate: npm run codex:gen-protocol
// Source file: v2/ToolRequestUserInputResponse.ts

import type { ToolRequestUserInputAnswer } from "./ToolRequestUserInputAnswer.js";

/**
 * EXPERIMENTAL. Response payload mapping question ids to answers.
 */
export type ToolRequestUserInputResponse = { answers: { [key in string]?: ToolRequestUserInputAnswer }, };